agent of justice phase-1/efforts/requirements.txt
flask
huggingface_hub
python-dotenv
flask-cors
